Default to debug level logs for unexpected verbosity
Before this change, a user specifying a verbosity > 3 would not hit any 
case in the verbosity switch statement, and the Go zero-valued default 
would result in FATAL level logs only. This is unlikely to be their 
intention.

This change will cause the log level to default to DEBUG, which is 
likely what a user specifying a high verbosity wants. A user who does 
not specify a default will still get WARN level logging by default due 
to CLI defaults.